learning-platform/pyproject.toml
target-version = "py313"

[project]
name = "learning-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-environ>=0.11.2",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "django-extensions>=3.2.3",
    "drf-spectacular>=0.28.0",
    "coverage>=7.6.12",
    "django-filter>=25.1",
    "factory-boy>=3.3.3",
    "pytest-django==4.5.2",
    "faker>=36.1.1",
    "celery>=5.5.2",
    "sendgrid>=6.12.0",
    "redis>=6.0.0",
    "split-settings>=1.0.0",
    "sentry-sdk>=2.26.1",
    "pytest-cov>=6.1.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=4.4.6",
    "pre-commit>=4.0.1",
    "ruff>=0.8.2",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = ["E501", "E741"]
